refactor: 입력, 출력을 담당하는 공통 함수 파일 이동
