--- src/utils/inputHandler.js
@@ -1,7 +1,0 @@
-import { MissionUtils } from "@woowacourse/mission-utils";
-
-const readUserInput = async (message) => {
-  return await MissionUtils.Console.readLineAsync(message);
-};
-
-export default readUserInput;
