--- src/utils/outputHandler.js
@@ -1,7 +1,0 @@
-import { MissionUtils } from "@woowacourse/mission-utils";
-
-const printResult = (message) => {
-  MissionUtils.Console.print(message);
-};
-
-export default printResult;
