+++ src/utils/util.js
@@ -1,6 +1,13 @@
+import { MissionUtils } from "@woowacourse/mission-utils";
+
+const readUserInput = async (message) =>
+  await MissionUtils.Console.readLineAsync(message);
+
+const printResult = (message) => MissionUtils.Console.print(message);
+
 const _pipe =
   (...fns) =>
   (input) =>
     fns.reduce((acc, fn) => fn(acc), input);
 
-export default _pipe;
+export { readUserInput, printResult, _pipe };
